Re-organize some code, change logger.warning to assertionFailure, reduce verbosity of extensions and file names
