+++ Sources/Utils/Data/Data+CommitmentCrc32.swift
@@ -7,8 +7,8 @@ import Foundation
 extension Data {
     // this function is only valid in the following case:
     //  - the Data instance is also a valid Data32
-    //  - the Data instance represents the data from a TxOutCommitment
-    var txOutCommitmentCrc32:UInt32? {
-        return Data32(self)?.txOutCommitmentCrc32
+    //  - the Data instance represents a CompressedCommitment
+    var commitmentCrc32: UInt32? {
+        Data32(self)?.commitmentCrc32
     }
 }
